backend/requirements.txt
ansi2html==1.8.0
antlr4-python3-runtime==4.13.0
antlr4-tools==0.2
argon2-cffi==20.1.0
asgiref==3.7.2
async-generator==1.10
attrs==21.2.0
autopep8==2.0.1
backcall==0.2.0
bleach==3.3.1
blinker==1.6.3
certifi==2021.10.8
cffi==1.14.6
charset-normalizer==2.0.7
click==8.1.3
cloudpickle==2.2.1
colorama==0.4.4
contourpy==1.0.7
crypto==1.4.1
cssselect==1.2.0
cycler==0.11.0
Cython==3.0.4
dash==2.14.1
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-table==5.0.0
debugpy==1.4.0
decorator==5.0.9
defusedxml==0.7.1
Django==4.1.10
django-cors-headers==4.2.0
djangorestframework==3.14.0
djongo==1.3.6
docutils==0.19
entrypoints==0.3
et-xmlfile==1.1.0
Farama-Notifications==0.0.4
filelock==3.12.0
Flask==3.0.0
fonttools==4.38.0
freetype-py==2.3.0
graphviz==0.20.1
greenlet==1.1.0
gymnasium==0.28.1
idna==3.3
imbalanced-learn==0.10.1
imblearn==0.0
importlib-metadata==6.0.0
install-jdk==1.1.0
ipykernel==6.0.3
ipython==7.25.0
ipython-genutils==0.2.0
ipython-sql==0.4.0
ipywidgets==7.6.3
itsdangerous==2.1.2
jaraco.classes==3.2.3
jax-jumpy==1.0.0
jedi==0.18.0
Jinja2==3.1.2
joblib==1.2.0
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==6.1.12
jupyter-console==6.4.0
jupyter-core==4.7.1
jupyterlab-pygments==0.1.2
jupyterlab-widgets==1.0.0
keyring==23.13.1
kiwisolver==1.4.4
leaf==1.0.7
lxml==4.9.2
markdown-it-py==2.1.0
MarkupSafe==2.1.3
matplotlib==3.6.3
matplotlib-inline==0.1.2
mdurl==0.1.2
mistune==0.8.4
more-itertools==9.0.0
mpmath==1.3.0
mtools==1.6.4
Naked==0.1.31
nbclient==0.5.3
nbconvert==6.1.0
nbformat==5.1.3
nest-asyncio==1.5.1
networkx==3.1
nltk==3.8.1
node==0.9.28
notebook==6.4.0
npm==0.1.1
numpy==1.26.1
odict==1.8.0
openpyxl==3.1.2
optional-django==0.1.0
packaging==23.1
pandas==1.5.3
pandocfilters==1.4.3
parso==0.8.2
patsy==0.5.3
pickleshare==0.7.5
Pillow==9.4.0
pkginfo==1.9.6
plotly==5.18.0
plumber==1.6
pmdarima==2.0.4
prettytable==0.7.2
prometheus-client==0.11.0
prompt-toolkit==3.0.19
psutil==5.8.0
psycopg2==2.9.2
pyagrum==1.5.2
pyasn1==0.4.8
pycairo==1.23.0
pycodestyle==2.10.0
pycparser==2.20
pydot==1.4.2
pygame==2.0.1
pygame-gui==0.5.7
pygame-menu==4.1.5
PyGLM==2.5.4
Pygments==2.14.0
pymongo==3.12.1
PyOpenGL==3.1.5
pyparsing==2.4.7
pyperclip==1.8.2
pyrsistent==0.18.0
python-dateutil==2.8.2
python-sql==1.2.2
pytz==2022.7.1
pywin32==301
pywin32-ctypes==0.2.0
pywinpty==1.1.3
PyYAML==6.0
pyzmq==22.1.0
qtconsole==5.1.1
QtPy==1.9.0
readme-renderer==37.3
RedesBayesianas-18364-18761==0.0.3
regex==2022.10.31
reportlab==4.0.0
requests==2.26.0
requests-toolbelt==0.10.1
retrying==1.3.4
rfc3986==2.0.0
rich==13.3.1
rlPyCairo==0.2.0
rsa==4.7.2
scikit-fuzzy==0.4.2
scikit-learn==1.2.1
scipy==1.10.0
seaborn==0.12.2
Send2Trash==1.7.1
shellescape==3.8.1
six==1.16.0
sql==0.4.0
SQLAlchemy==1.4.22
sqlparse==0.2.4
statsmodels==0.13.5
sympy==1.12
tabulate==0.8.9
tenacity==8.2.3
terminado==0.10.1
testpath==0.5.0
threadpoolctl==3.1.0
tomli==2.0.1
torch==2.0.1
torchvision==0.15.2
tornado==6.1
tqdm==4.64.1
traitlets==5.0.5
twine==4.0.2
typing_extensions==4.5.0
tzdata==2023.3
urllib3==1.26.7
wcwidth==0.2.5
webencodings==0.5.1
Werkzeug==3.0.1
widgetsnbextension==3.5.1
wordcloud==1.8.2.2
zipp==3.12.1
zope.component==5.0.1
zope.deprecation==4.4.0
zope.event==4.5.0
zope.hookable==5.1.0
zope.interface==5.4.0
zope.lifecycleevent==4.3